+++ src/index.js
@@ -1,9 +1,33 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { createStore } from 'redux';
 import { Provider } from 'react-redux';
+
 import App from './components/App';
-import store from './reducers/index';
+import rootReducer from './reducers/index';
 import './style.css';
+
+const initialState = {
+  books: [
+    {
+      id: Math.random(),
+      title: 'A Fiction Book',
+      category: 'Sci-Fi',
+    },
+    {
+      id: Math.random(),
+      title: 'Some guy bio',
+      category: 'Biography',
+    },
+    {
+      id: Math.random(),
+      title: 'Watch out, the return',
+      category: 'Horror',
+    },
+  ],
+};
+
+const store = createStore(rootReducer, initialState);
 
 ReactDOM.render(
   <Provider store={store}>
@@ -11,3 +35,5 @@ ReactDOM.render(
   </Provider>,
   document.getElementById('root'),
 );
+
+export default initialState;
